op3: prop: Drive UI thread runnable time to zero
Source: https://source.android.com/devices/tech/debug/jank_jitter
Commit: https://android.googlesource.com/platform/frameworks/base/+/33eb07f

"Warning: Do not use this option on heterogeneous CPU configurations
unless you have a capacity-aware RT scheduler. And, at this moment, NO
CURRENTLY SHIPPING RT SCHEDULER IS CAPACITY-AWARE. We're working on one
for EAS, but it's not yet available. The default RT scheduler is based
purely on RT priorities and whether a CPU already has a RT thread of
equal or higher priority.

As a result, the default RT scheduler will happily move your
relatively-long-running UI thread from a high-frequency big core to a
little core at minimum frequency if a higher priority FIFO kthread
happens to wake up on the same big core. This will introduce significant
performance regressions. As this option has not yet been used on a
shipping Android device, if you want to use it get in touch with the
Android performance team to help you validate it.

When sys.use_fifo_ui is enabled, ActivityManager tracks the UI thread
and RenderThread (the two most UI-critical threads) of the top
application and makes those threads SCHED_FIFO instead of SCHED_OTHER.
This effectively eliminates jitter from UI and RenderThreads; the traces
we've gathered with this option enabled show runnable times on the order
of microseconds instead of milliseconds.

However, because the RT load balancer was not capacity-aware, there was
a 30% reduction in application startup performance because the UI thread
responsible for starting up the app would be moved from a 2.1Ghz gold
Kryo core to a 1.5GHz silver Kryo core. With a capacity-aware RT load
balancer, we see equivalent performance in bulk operations and a 10-15%
reduction in 95th and 99th percentile frame times in many of our UI
benchmarks."

Signed-off-by: joshuous <joshuous@gmail.com>